requirements.txt
async-timeout==4.0.2
asynctest==0.13.0
beautifulsoup4==4.10.0
bookmarks-parser==0.1.3
bs4==0.0.1
faiss-cpu==1.7.2
google-auth==2.5.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
h5py==3.6.0
html-to-json==2.0.0
html2text==2020.1.16
html5lib==1.1
huggingface==0.0.1
huggingface-hub==0.4.0
idna==3.3
imageio==2.14.1
leveldb==0.201
mailbox==0.4
mammoth==1.4.18
markdown==3.3.6
markdownify==0.10.1
newspaper3k==0.2.8
nltk==3.6.5
numpy==1.21.5
pandas==1.3.5
pdfminer.six==20221105
pillow==8.4.0
pylev==1.4.0
pyyaml==6.0
rake-nltk==1.0.6
readability-lxml==0.8.1
scikit-learn==1.0.2
scipy==1.7.3
selectivesearch==0.4
sentence-transformers==2.2.0
spacy==3.2.1
tqdm==4.62.3
transformers==4.16.1
urllib3==1.26.7
validators==0.18.2
wand==0.6.7
wasabi==0.9.0
yake==0.4.8